requirements.txt
google-generativeai
python-dotenv
Jinja2
premailer
markdown2
